Carrinho de compras funcionando

+++ NossaCasaDoCodigo.Testes/TestaLivro.cs
@@ -18,7 +18,7 @@ namespace NossaCasaDoCodigo.Testes
         [Theory]
         [MemberData(nameof(TeoriasInvalidasTitulo))]
         public void TituloNaoPodeSerVazioOuNulo(string titulo, string resumo, string sumario, int paginas,
-            string iSBN, Autor autor, Categoria categoria, double edicao, double preco)
+            string iSBN, Autor autor, Categoria categoria, double edicao, decimal preco)
         {
             Assert.Throws<ArgumentNullException>(() => new Livro(titulo, resumo, sumario, paginas,
                 iSBN, autor, categoria, edicao, preco));
@@ -31,7 +31,7 @@ namespace NossaCasaDoCodigo.Testes
 
             var novoLivro = new Livro("PHP para baixinhos",
                 LivrosHelpers.RESUMO_COM_MAIS_DE_500_CARACTERES, "Teste", 20, "978-12-34567-89-0",
-                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.0, 35.50);
+                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.0, 35.50m);
 
             Assert.True(LivrosDAO.BuscarTitulo(novoLivro.Titulo).Any());
         }
@@ -41,13 +41,13 @@ namespace NossaCasaDoCodigo.Testes
         {
             Assert.Throws<ArgumentException>(() => new Livro("Titulo Novo", "Bem menos de 500 caracteres.",
                 null, 20, "978-12-34567-89-0",
-                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.1, 50.95));
+                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.1, 50.95m));
         }
 
         [Theory]
         [MemberData(nameof(TeoriasInvalidasResumo))]
         public void ResumoNaoPodeSerVazioOuNulo(string titulo, string resumo, string sumario, int paginas,
-            string iSBN, Autor autor, Categoria categoria, double edicao, double preco)
+            string iSBN, Autor autor, Categoria categoria, double edicao, decimal preco)
         {
             Assert.Throws<ArgumentNullException>(() => new Livro(titulo, resumo, sumario, paginas,
                 iSBN, autor, categoria, edicao, preco));
@@ -56,7 +56,7 @@ namespace NossaCasaDoCodigo.Testes
         [Theory]
         [MemberData(nameof(TeoriasInvalidasSumario))]
         public void SumarioNaoPodeSerVazioOuNulo(string titulo, string resumo, string sumario, int paginas,
-            string iSBN, Autor autor, Categoria categoria, double edicao, double preco)
+            string iSBN, Autor autor, Categoria categoria, double edicao, decimal preco)
         {
             Assert.Throws<ArgumentNullException>(() => new Livro(titulo, resumo, sumario, paginas,
                 iSBN, autor, categoria, edicao, preco));
@@ -65,7 +65,7 @@ namespace NossaCasaDoCodigo.Testes
         [Theory]
         [MemberData(nameof(TeoriasInvalidasISBN))]
         public void IsbnNaoPodeTerFormatoInvalido(string titulo, string resumo, string sumario, int paginas,
-            string iSBN, Autor autor, Categoria categoria, double edicao, double preco)
+            string iSBN, Autor autor, Categoria categoria, double edicao, decimal preco)
         {
             Assert.Throws<ArgumentException>(() => new Livro(titulo, resumo, sumario, paginas,
                 iSBN, autor, categoria, edicao, preco));
@@ -77,7 +77,7 @@ namespace NossaCasaDoCodigo.Testes
         {
             Assert.Throws<ArgumentException>(() => new Livro("teste",
                 LivrosHelpers.RESUMO_COM_MAIS_DE_500_CARACTERES, "Teste", 20, "978-12-34567-89-0",
-                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 0, 50.95));
+                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 0, 50.95m));
         }
 
         [Fact]
@@ -93,13 +93,13 @@ namespace NossaCasaDoCodigo.Testes
             new LivrosDAO();
             LivrosDAO.Salvar(new Livro("PHP para baixinhos",
                 LivrosHelpers.RESUMO_COM_MAIS_DE_500_CARACTERES, "Teste", 20, "978-12-34567-89-0",
-                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.0, 35.50));
+                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.0, 35.50m));
             LivrosDAO.Salvar(new Livro("Escrevi este livro e fiquei rico!",
                 LivrosHelpers.RESUMO_COM_MAIS_DE_500_CARACTERES, "Teste", 20, "978-12-34567-89-0",
-                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.0, 35.50));
+                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.0, 35.50m));
             LivrosDAO.Salvar(new Livro("Seja um coach quântico sem saber o significado de quântico.",
                 LivrosHelpers.RESUMO_COM_MAIS_DE_500_CARACTERES, "Teste", 20, "978-12-34567-89-0",
-                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.0, 35.50));
+                new Autor("nome", "email@gmail.com", "minha descrição"), new Categoria("categoria"), 2.0, 35.50m));
         }
     }
 }
